refactor(metrics): using seperate module for runtime metrics
